Refactor: [#90] DTO 패키지 내 VO 분리
