+++ BE/src/main/java/com/enolj/airbnb/web/vo/Charge.java
@@ -1,4 +1,4 @@
-package com.enolj.airbnb.web.dto;
+package com.enolj.airbnb.web.vo;
 
 public class Charge {
